pyproject.toml
[project]
name = "realsense-camera-calibration"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.4.3",
    "opencv-python>=4.13.0.92",
    "pillow>=12.1.1",
    "pyrealsense2>=2.57.7.10387",
    "pyyaml>=6.0",
]
